Move CLI code into commands module
